# [NOSQUASH] Remove the unnecessary sort in CNullDriver::addTexture