@sfan5 2023-09-13
I don't think we use this but this sort call can be removed too:
![grafik](https://github.com/minetest/irrlicht/assets/1042418/877646f9-7d65-43d8-a563-52596181f425)