@Desour 2023-09-13
`renameTexture` is actually broken, if I'm not wrong, because `Textures` is not flagged as unsorted before the `Textures.sort()` call.
I've removed the function now, together with another unused not useful one.