@appgurueu 2023-09-13
(For the future) could this be refactored to use an unordered map?