@sfan5 2023-09-13
As I just found out [here] texture names are apparently not an one-to-one mapping to textures :thinking: 
So, maybe.


[here]: https://irc.minetest.net/minetest-dev/2023-09-13#i_6113112