@Desour 2023-09-13
If they are not supposed to be unique, we could still use an unordered multimap.
It would make sense imo, I just didn't do it because I wanted the change to be simple. (Also, if I change the header, I have to recompile half of minetest.)