@SmallJoker 2024-02-10
The Irrlicht situation, visualized:
![xkcd Standards](https://imgs.xkcd.com/comics/standards.png)

 @BeverlyCode Please feel free to forward such patches to upstream Irrlicht so that all forks *could* benefit from it. Fixes in private forks do not help anyone else.