@sfan5 2024-02-15
Our primary concern are security-relevant issues in the engine or IrrlichtMt.
You can disclose them as written in our [security policy](https://github.com/minetest/minetest/security/policy) and they will absolutely get looked at.

> and to be honest I don't really like any of you. 🤷

That is regrettable but it was your choice to comment here.